Bug 1998365 - Switch NonCustomCSSPropertyId to uint16_t; r=emilio,firefox-style-system-reviewers
This patch changes NonCustomCSSPropertyId to use uint16_t as its underlying
type, matching the Rust representation.

Because the enum is now unsigned, the UNKNOWN sentinel can no longer be -1.
Keeping UNKNOWN as 0 would shift all property ids and complicate existing
mappings, so UNKNOWN is moved to the end of the enum instead.

The patch also updates the IPC serialization logic. Previously it assumed that
_COUNT covered almost every property, but over time additional entries (such
as aliases) were added after _COUNT and the IPC handling was not updated. This
patch corrects the serialization so that all variants of the enum are covered,
in line with the original intent of bug 783835.

No behavior changes intended beyond fixing the serialization gap.

Differential Revision: https://phabricator.services.mozilla.com/D271566

UltraBlame original commit: 4e768c9d7c484592504a23621d6df0dade2bb53b

+++ dom/html/nsGenericHTMLElement.h
@@ -28,7 +28,7 @@ class nsILayoutHistoryState;
 class nsIURI;
 struct nsSize;
 
-enum NonCustomCSSPropertyId : int32_t;
+enum NonCustomCSSPropertyId : uint16_t;
 
 namespace mozilla {
 class EditorBase;
